Decouple UI from models + map code; kick jQuery habit
Everything was tightly coupled, and heavily reliant on jQuery in the
process. Refactored a lot of the UI code out to separate classes in
the ui/ folder, and removed all jQuery dependencies from code not in
ui/ as much as possible.
(OrbitControls still TODO, but that one needs more work in general)