Allow XY pan with 4 fingers or alt
Camera is now less stuck to the grid, and you can move it up and down by
swiping with four fingers, or holding alt while right click and
dragging.

Also fixed a minor bug where ctrl-l would toggle locking, which is
annoying when you want to just to the URL of the browser.